Move `contains_return` to utils/mod.rs
